Refactored CLI plugins
