+++ tools/plugins/preinstall/src/main/kotlin/net/corda/cli/plugins/preinstall/CheckKafka.kt
@@ -1,0 +1,2 @@
+package net.corda.cli.plugins.preinstall
+
